refactor: resetting spring boot project
