+++ BE/settings.gradle.kts
@@ -1,1 +1,1 @@
-rootProject.name = "BE"
+rootProject.name = "be"
