+++ BE/src/main/java/ssafy/antalbum/be/BeApplication.java
@@ -1,4 +1,4 @@
-package com.ssafy.antalbum;
+package ssafy.antalbum.be;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class BeApplication {
 
-    public static void main(String[] args) {
-        SpringApplication.run(BeApplication.class, args);
-    }
+	public static void main(String[] args) {
+		SpringApplication.run(BeApplication.class, args);
+	}
 
 }
