+++ BE/src/main/java/ssafy/antalbum/be/Hello.java
@@ -1,0 +1,10 @@
+package ssafy.antalbum.be;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class Hello {
+    private String data;
+}
