+++ BE/src/main/java/ssafy/antalbum/be/HelloController.java
@@ -1,0 +1,14 @@
+package ssafy.antalbum.be;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class HelloController {
+
+    @GetMapping("hello")
+    public String hello() {
+        return "hello";
+    }
+
+}
