--- BE/src/main/resources/application.properties
@@ -1,8 +1,0 @@
-# DataSource
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.datasource.url=jdbc:mysql://localhost:3306/antalbum?useSSL=false&useUnicode=true&serverTimezone=Asia/Seoul
-spring.datasource.username=antalbum
-spring.datasource.password=ssafy
-spring.jpa.show-sql=true
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.properties.hibernate.format_sql=true
