+++ BE/src/main/resources/spy.properties
@@ -1,0 +1,2 @@
+appender=com.p6spy.engine.spy.appender.Slf4JLogger
+logMessageFormat=com.p6spy.engine.spy.appender.MultiLineFormat
