+++ BE/src/test/java/ssafy/antalbum/be/BeApplicationTests.java
@@ -1,4 +1,4 @@
-package com.ssafy.antalbum;
+package ssafy.antalbum.be;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -6,8 +6,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 class BeApplicationTests {
 
-    @Test
-    void contextLoads() {
-    }
+	@Test
+	void contextLoads() {
+	}
 
 }
